feat(legend): add color picker state for legend item
